[Chore] #56 - Service 코드 파일 분리 및 코드 자잘한 수정
