+++ Projects/Data/RecommendationData/Sources/Service/DefaultRecommendationService.swift
@@ -1,0 +1,54 @@
+//
+//  DefaultRecommendationService.swift
+//  RecommendationData
+//
+//  Created by Seoyeon Choi on 4/26/26.
+//  Copyright © 2026 kr.websoso.app. All rights reserved.
+//
+
+import Foundation
+import Networking
+
+public struct DefaultRecommendationService: RecommendationService {
+    
+    private let network: NetworkingRequestable
+    
+    public init(network: NetworkingRequestable) {
+        self.network = network
+    }
+    
+    public func getTodayDiscovery() async throws -> TodayDiscoveryNovelsResponse {
+        try await network.request(
+            RecommendationEndpoint.getTodayDiscovery,
+            decodeTo: TodayDiscoveryNovelsResponse.self
+        )
+    }
+    
+    public func getTrendingFeeds() async throws -> TrendingFeedsResponse {
+        try await network.request(
+            RecommendationEndpoint.getTrendingFeeds,
+            decodeTo: TrendingFeedsResponse.self
+        )
+    }
+    
+    public func getInterestFeeds() async throws -> InterestFeedsResponse {
+        try await network.request(
+            RecommendationEndpoint.getInterestFeeds,
+            decodeTo: InterestFeedsResponse.self
+        )
+    }
+    
+    public func getPreferenceGenreNovels() async throws -> PreferenceGenreNovelsResponse {
+        try await network.request(
+            RecommendationEndpoint.getPreferenceGenreNovels,
+            decodeTo: PreferenceGenreNovelsResponse.self
+        )
+    }
+    
+    public func getSosopickNovels() async throws -> SosopickNovelsResponse {
+        try await network.request(
+            RecommendationEndpoint.sosopickNovels,
+            decodeTo: SosopickNovelsResponse.self
+        )
+    }
+}
